+++ NEWS.md
@@ -5,6 +5,8 @@ skimr 1.0.1 (2018-01-xx)
 
 ### MAJOR CHANGES
    * spark_line() and spark_bar() are no longer exported
+   * Default statistics for numeric changed from min(x) and max(x) to 
+     quantile(x, probs= c(0)) and quantile(x, probs= c(1))
 
 #### MINOR IMPROVEMENTS
    * Add minimimum required version for stringr